pyproject.toml
[project]
name = "workload-identity-oauth-reverse-proxy"
version = "0.1.0"
description = "Proof of Concept for issuing droplets a workload identity token based on scopes on droplet create. Token gives access to spaces key creation and databases based on droplet tags."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "uv",
  "pyjwt",
  "jwcrypto",
  "jsonschema",
  "cryptography",
  "pydantic[email]",
  "paramiko",
  "pyyaml>=6.0.2",
  "python-hcl2>=7.3.1",
  "psycopg2-binary>=2.9.10",
  "sqlalchemy>=2.0.43",
  "snoop>=0.6.0",
  "markdown>=3.9",
  "pygments>=2.19.1",
  "beautifulsoup4>=4.13.5",
  "lxml>=6.0.1",
  "aiohttp>=3.12.15",
  "atproto>=0.0.62",
  "pydantic>=2.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/workload_identity_oauth_reverse_proxy"]
